requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bleak==1.1.1
certifi==2025.11.12
charset-normalizer==3.4.4
dbus-fast==3.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
meshtastic==2.7.5
ollama==0.6.1
packaging==24.2
protobuf==6.33.1
pydantic==2.12.4
pydantic_core==2.41.5
Pypubsub==4.0.3
pyserial==3.5
PyYAML==6.0.3
requests==2.32.5
sniffio==1.3.1
tabulate==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
